changed parameters in view to work with controller

+++ src/View/CellVisual.java
@@ -4,11 +4,11 @@ import javafx.scene.paint.Paint;
 import javafx.scene.shape.Rectangle;
 
 public class CellVisual extends Rectangle {
-    double myXPos;
-    double myYPos;
+    int myXPos;
+    int myYPos;
     Paint myColor;
 
-    public CellVisual(double width, double height, Paint color, double x, double y) {
+    public CellVisual(double width, double height, Paint color, int x, int y) {
         this.myXPos = x;
         this.myYPos = y;
         this.myColor = color;
@@ -16,10 +16,10 @@ public class CellVisual extends Rectangle {
         this.setHeight(height);
         this.setFill(color);
     }
-    public double getXPos() {
+    public int getXPos() {
         return myXPos;
     }
-    public double getYPos() {
+    public int getYPos() {
         return myYPos;
     }
 }
